# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.ConnectorCreateRequest:
    properties:
      display_name:
        type: string
      exchange_id:
        type: string
      rate_limit:
        properties:
          limit:
            description: Max requests per period
            minimum: 1
            type: integer
          min_delay_ms:
            description: 'Min delay between calls (default: calculated from limit/period)'
            minimum: 100
            type: integer
          period_ms:
            description: Period in milliseconds
            minimum: 1000
            type: integer
        required:
        - limit
        - period_ms
        type: object
    required:
    - display_name
    - exchange_id
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@datacollector.io
    name: API Support
    url: https://github.com/yourusername/datacollector
  description: Cryptocurrency market data collection service with OHLCV data, technical
    indicators, and job management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DataCollector API
  version: 1.0.5
paths:
  /connectors:
    get:
      consumes:
      - application/json
      description: Retrieves all exchange connectors with job counts
      parameters:
      - description: Filter by status (active, suspended)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of connectors
          schema:
            additionalProperties: true
            type: object
      summary: Get all connectors
      tags:
      - Connectors
    post:
      consumes:
      - application/json
      description: Creates a new exchange connector with rate limiting configuration
      parameters:
      - description: Connector configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ConnectorCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Connector created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Connector already exists
          schema:
            additionalProperties: true
            type: object
      summary: Create a new connector
      tags:
      - Connectors
...
schemes:
- http
- https
swagger: "2.0"